refactor: cleaner layout with collapsible view menu options
