+++ apps/readest-app/public/locales/ar/translation.json
@@ -1060,7 +1060,5 @@
   "Library Sync Progress": "تقدم تزامن المكتبة",
   "Back to library": "العودة إلى المكتبة",
   "Manual": "يدوي",
-  "Group by...": "تجميع حسب...",
-  "Group By": "تجميع حسب",
-  "Sort": "فرز"
+  "Group by...": "تجميع حسب..."
 }
