+++ apps/readest-app/public/locales/bn/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "লাইব্রেরি সিঙ্ক অগ্রগতি",
   "Back to library": "লাইব্রেরিতে ফিরে যান",
   "Manual": "ম্যানুয়াল",
-  "Group by...": "গ্রুপ অনুযায়ী...",
-  "Group By": "গ্রুপ অনুযায়ী",
-  "Sort": "সাজান"
+  "Group by...": "গ্রুপ অনুযায়ী..."
 }
