+++ apps/readest-app/public/locales/bo/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "དཔེ་མཛོད་མཉམ་འབྱུང་གི་རིམ་པ།",
   "Back to library": "དཔེ་མཛོད་ལ་ལོག་པ།",
   "Manual": "ལག་བཟོས།",
-  "Group by...": "དབྱེ་བ་འབྱེད་སྟངས།...",
-  "Group By": "དབྱེ་བ་འབྱེད་སྟངས།",
-  "Sort": "རིམ་སྒྲིག་བྱེད་པ།"
+  "Group by...": "དབྱེ་བ་འབྱེད་སྟངས།..."
 }
