+++ apps/readest-app/public/locales/el/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "Πρόοδος συγχρονισμού βιβλιοθήκης",
   "Back to library": "Επιστροφή στη βιβλιοθήκη",
   "Manual": "Χειροκίνητα",
-  "Group by...": "Ομαδοποίηση κατά...",
-  "Group By": "Ομαδοποίηση κατά",
-  "Sort": "Ταξινόμηση"
+  "Group by...": "Ομαδοποίηση κατά..."
 }
