+++ apps/readest-app/public/locales/es/translation.json
@@ -1024,7 +1024,5 @@
   "Library Sync Progress": "Progreso de sincronización de la biblioteca",
   "Back to library": "Volver a la biblioteca",
   "Manual": "Manual",
-  "Group by...": "Agrupar por...",
-  "Group By": "Agrupar por",
-  "Sort": "Ordenar"
+  "Group by...": "Agrupar por..."
 }
