+++ apps/readest-app/public/locales/fa/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "پیشرفت همگام‌سازی کتابخانه",
   "Back to library": "بازگشت به کتابخانه",
   "Manual": "دستی",
-  "Group by...": "گروه‌بندی بر اساس...",
-  "Group By": "گروه‌بندی بر اساس",
-  "Sort": "مرتب‌سازی"
+  "Group by...": "گروه‌بندی بر اساس..."
 }
