+++ apps/readest-app/public/locales/fr/translation.json
@@ -1024,7 +1024,5 @@
   "Library Sync Progress": "Progression de la synchronisation de la bibliothèque",
   "Back to library": "Retour à la bibliothèque",
   "Manual": "Manuel",
-  "Group by...": "Grouper par...",
-  "Group By": "Grouper par",
-  "Sort": "Trier"
+  "Group by...": "Grouper par..."
 }
