+++ apps/readest-app/public/locales/hi/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "पुस्तकालय सिंक प्रगति",
   "Back to library": "लाइब्रेरी पर वापस जाएं",
   "Manual": "मैनुअल",
-  "Group by...": "इसके द्वारा समूहबद्ध करें...",
-  "Group By": "इसके द्वारा समूहबद्ध करें",
-  "Sort": "क्रमबद्ध करें"
+  "Group by...": "इसके द्वारा समूहबद्ध करें..."
 }
