+++ apps/readest-app/public/locales/id/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "Kemajuan Sinkronisasi Pustaka",
   "Back to library": "Kembali ke perpustakaan",
   "Manual": "Manual",
-  "Group by...": "Kelompokkan menurut...",
-  "Group By": "Kelompokkan menurut",
-  "Sort": "Urutkan"
+  "Group by...": "Kelompokkan menurut..."
 }
