+++ apps/readest-app/public/locales/it/translation.json
@@ -1024,7 +1024,5 @@
   "Library Sync Progress": "Avanzamento sincronizzazione libreria",
   "Back to library": "Torna alla libreria",
   "Manual": "Manuale",
-  "Group by...": "Raggruppa per...",
-  "Group By": "Raggruppa per",
-  "Sort": "Ordina"
+  "Group by...": "Raggruppa per..."
 }
