+++ apps/readest-app/public/locales/ja/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "ライブラリ同期の進捗",
   "Back to library": "ライブラリに戻る",
   "Manual": "手動",
-  "Group by...": "グループ化...",
-  "Group By": "グループ化",
-  "Sort": "並べ替え"
+  "Group by...": "グループ化..."
 }
