+++ apps/readest-app/public/locales/ko/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "라이브러리 동기화 진행 상황",
   "Back to library": "라이브러리로 돌아가기",
   "Manual": "수동",
-  "Group by...": "그룹화 기준...",
-  "Group By": "그룹화 기준",
-  "Sort": "정렬"
+  "Group by...": "그룹화 기준..."
 }
