+++ apps/readest-app/public/locales/ms/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "Kemajuan Segerak Perpustakaan",
   "Back to library": "Kembali ke perpustakaan",
   "Manual": "Manual",
-  "Group by...": "Kumpulan mengikut...",
-  "Group By": "Kumpulan mengikut",
-  "Sort": "Susun"
+  "Group by...": "Kumpulan mengikut..."
 }
