+++ apps/readest-app/public/locales/nl/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "Voortgang bibliotheeksynchronisatie",
   "Back to library": "Terug naar bibliotheek",
   "Manual": "Handmatig",
-  "Group by...": "Groeperen op...",
-  "Group By": "Groeperen op",
-  "Sort": "Sorteren"
+  "Group by...": "Groeperen op..."
 }
