+++ apps/readest-app/public/locales/pl/translation.json
@@ -1036,7 +1036,5 @@
   "Library Sync Progress": "Postęp synchronizacji biblioteki",
   "Back to library": "Powrót do biblioteki",
   "Manual": "Ręcznie",
-  "Group by...": "Grupuj według...",
-  "Group By": "Grupuj według",
-  "Sort": "Sortuj"
+  "Group by...": "Grupuj według..."
 }
