+++ apps/readest-app/public/locales/pt/translation.json
@@ -1024,7 +1024,5 @@
   "Library Sync Progress": "Progresso de sincronização da biblioteca",
   "Back to library": "Voltar à biblioteca",
   "Manual": "Manual",
-  "Group by...": "Agrupar por...",
-  "Group By": "Agrupar por",
-  "Sort": "Ordenar"
+  "Group by...": "Agrupar por..."
 }
