+++ apps/readest-app/public/locales/ru/translation.json
@@ -1036,7 +1036,5 @@
   "Library Sync Progress": "Прогресс синхронизации библиотеки",
   "Back to library": "Назад в библиотеку",
   "Manual": "Вручную",
-  "Group by...": "Группировать по...",
-  "Group By": "Группировать по",
-  "Sort": "Сортировать"
+  "Group by...": "Группировать по..."
 }
