+++ apps/readest-app/public/locales/si/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "පුස්තකාල සමමුහුර්ත කිරීමේ ප්‍රගතිය",
   "Back to library": "පුස්තකාලයට ආපසු",
   "Manual": "අත්පොත",
-  "Group by...": "මගින් සමූහනය කරන්න...",
-  "Group By": "මගින් සමූහනය කරන්න",
-  "Sort": "වර්ග කරන්න"
+  "Group by...": "මගින් සමූහනය කරන්න..."
 }
