+++ apps/readest-app/public/locales/sv/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "Synkroniseringsframsteg för bibliotek",
   "Back to library": "Tillbaka till biblioteket",
   "Manual": "Manuellt",
-  "Group by...": "Gruppera efter...",
-  "Group By": "Gruppera efter",
-  "Sort": "Sortera"
+  "Group by...": "Gruppera efter..."
 }
