+++ apps/readest-app/public/locales/ta/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "நூலக ஒத்திசைவு நிலை",
   "Back to library": "நூலகத்திற்குத் திரும்பு",
   "Manual": "கையேடு",
-  "Group by...": "இதன் படி குழுவாக்கு...",
-  "Group By": "இதன் படி குழுவாக்கு",
-  "Sort": "வரிசைப்படுத்து"
+  "Group by...": "இதன் படி குழுவாக்கு..."
 }
