+++ apps/readest-app/public/locales/th/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "ความคืบหน้าการซิงค์ห้องสมุด",
   "Back to library": "กลับไปที่ห้องสมุด",
   "Manual": "ทำเอง",
-  "Group by...": "จัดกลุ่มตาม...",
-  "Group By": "จัดกลุ่มตาม",
-  "Sort": "เรียงลำดับ"
+  "Group by...": "จัดกลุ่มตาม..."
 }
