+++ apps/readest-app/public/locales/tr/translation.json
@@ -1012,7 +1012,5 @@
   "Library Sync Progress": "Kütüphane Senkronizasyon İlerlemesi",
   "Back to library": "Kütüphaneye dön",
   "Manual": "Manuel",
-  "Group by...": "Şuna göre grupla...",
-  "Group By": "Şuna göre grupla",
-  "Sort": "Sırala"
+  "Group by...": "Şuna göre grupla..."
 }
