+++ apps/readest-app/public/locales/uk/translation.json
@@ -1036,7 +1036,5 @@
   "Library Sync Progress": "Прогрес синхронізації бібліотеки",
   "Back to library": "Назад до бібліотеки",
   "Manual": "Вручну",
-  "Group by...": "Групувати за...",
-  "Group By": "Групувати за",
-  "Sort": "Сортувати"
+  "Group by...": "Групувати за..."
 }
