+++ apps/readest-app/public/locales/vi/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "Tiến trình đồng bộ hóa thư viện",
   "Back to library": "Quay lại thư viện",
   "Manual": "Thủ công",
-  "Group by...": "Nhóm theo...",
-  "Group By": "Nhóm theo",
-  "Sort": "Sắp xếp"
+  "Group by...": "Nhóm theo..."
 }
