+++ apps/readest-app/public/locales/zh-CN/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "库同步进度",
   "Back to library": "返回书库",
   "Manual": "手动",
-  "Group by...": "分组方式...",
-  "Group By": "分组方式",
-  "Sort": "排序"
+  "Group by...": "分组方式..."
 }
