+++ apps/readest-app/public/locales/zh-TW/translation.json
@@ -1000,7 +1000,5 @@
   "Library Sync Progress": "資料庫同步進度",
   "Back to library": "返回書庫",
   "Manual": "手動",
-  "Group by...": "分組方式...",
-  "Group By": "分組方式",
-  "Sort": "排序"
+  "Group by...": "分組方式..."
 }
