+++ apps/readest-app/src/app/library/components/ViewMenu.tsx
@@ -3,7 +3,12 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { useEnv } from '@/context/EnvContext';
 import { useSettingsStore } from '@/store/settingsStore';
 import { useTranslation } from '@/hooks/useTranslation';
-import { LibraryCoverFitType, LibraryViewModeType } from '@/types/settings';
+import {
+  LibraryCoverFitType,
+  LibraryViewModeType,
+  LibraryGroupByType,
+  LibrarySortByType,
+} from '@/types/settings';
 import { saveSysSettings } from '@/helpers/settings';
 import { navigateToLibrary } from '@/utils/nav';
 import NumberInput from '@/components/settings/NumberInput';
@@ -25,6 +30,9 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
   const coverFit = settings.libraryCoverFit;
   const autoColumns = settings.libraryAutoColumns;
   const columns = settings.libraryColumns;
+  const groupBy = settings.libraryGroupBy;
+  const sortBy = settings.librarySortBy;
+  const isAscending = settings.librarySortAscending;
 
   const viewOptions = [
     { label: _('List'), value: 'list' },
@@ -34,6 +42,27 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
   const coverFitOptions = [
     { label: _('Crop'), value: 'crop' },
     { label: _('Fit'), value: 'fit' },
+  ];
+
+  const groupByOptions = [
+    { label: _('None'), value: LibraryGroupByType.None },
+    { label: _('Manual'), value: LibraryGroupByType.Manual },
+    { label: _('Series'), value: LibraryGroupByType.Series },
+    { label: _('Author'), value: LibraryGroupByType.Author },
+  ];
+
+  const sortByOptions = [
+    { label: _('Title'), value: LibrarySortByType.Title },
+    { label: _('Author'), value: LibrarySortByType.Author },
+    { label: _('Format'), value: LibrarySortByType.Format },
+    { label: _('Date Read'), value: LibrarySortByType.Updated },
+    { label: _('Date Added'), value: LibrarySortByType.Created },
+    { label: _('Date Published'), value: LibrarySortByType.Published },
+  ];
+
+  const sortingOptions = [
+    { label: _('Ascending'), value: true },
+    { label: _('Descending'), value: false },
   ];
 
   const handleSetViewMode = async (value: LibraryViewModeType) => {
@@ -62,11 +91,42 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
     await saveSysSettings(envConfig, 'libraryAutoColumns', false);
   };
 
+  const handleSetGroupBy = async (value: LibraryGroupByType) => {
+    await saveSysSettings(envConfig, 'libraryGroupBy', value);
+
+    const params = new URLSearchParams(searchParams?.toString());
+    if (value === LibraryGroupByType.Manual) {
+      params.delete('groupBy');
+    } else {
+      params.set('groupBy', value);
+    }
+    // Clear group navigation when changing groupBy mode
+    params.delete('group');
+    navigateToLibrary(router, `${params.toString()}`);
+  };
+
+  const handleSetSortBy = async (value: LibrarySortByType) => {
+    await saveSysSettings(envConfig, 'librarySortBy', value);
+
+    const params = new URLSearchParams(searchParams?.toString());
+    params.set('sort', value);
+    navigateToLibrary(router, `${params.toString()}`);
+  };
+
+  const handleSetSortAscending = async (value: boolean) => {
+    await saveSysSettings(envConfig, 'librarySortAscending', value);
+
+    const params = new URLSearchParams(searchParams?.toString());
+    params.set('order', value ? 'asc' : 'desc');
+    navigateToLibrary(router, `${params.toString()}`);
+  };
+
   return (
     <Menu
       className='view-menu dropdown-content no-triangle z-20 mt-2 shadow-2xl'
       onCancel={() => setIsDropdownOpen?.(false)}
     >
+      {/* View Mode */}
       {viewOptions.map((option) => (
         <MenuItem
           key={option.value}
@@ -77,6 +137,8 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
           transient
         />
       ))}
+
+      {/* Columns */}
       <hr aria-hidden='true' className='border-base-200 my-1' />
       <MenuItem label={_('Columns')} buttonClass='h-8' labelClass='text-sm sm:text-xs' disabled />
       <MenuItem
@@ -98,6 +160,8 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
         }
         onClick={() => handleToggleAutoColumns()}
       />
+
+      {/* Book Covers */}
       <hr aria-hidden='true' className='border-base-200 my-1' />
       <MenuItem
         label={_('Book Covers')}
@@ -115,6 +179,51 @@ const ViewMenu: React.FC<ViewMenuProps> = ({ setIsDropdownOpen }) => {
           transient
         />
       ))}
+
+      {/* Group By - Collapsible */}
+      <hr aria-hidden='true' className='border-base-200 my-1' />
+      <MenuItem label={_('Group by...')} detailsOpen={true} buttonClass='py-[4px]'>
+        <ul className='ms-0 flex flex-col ps-0 before:hidden'>
+          {groupByOptions.map((option) => (
+            <MenuItem
+              key={option.value}
+              label={option.label}
+              buttonClass='h-8'
+              toggled={groupBy === option.value}
+              onClick={() => handleSetGroupBy(option.value as LibraryGroupByType)}
+              transient
+            />
+          ))}
+        </ul>
+      </MenuItem>
+
+      {/* Sort By - Collapsible */}
+      <hr aria-hidden='true' className='border-base-200 my-1' />
+      <MenuItem label={_('Sort by...')} detailsOpen={false} buttonClass='py-[4px]'>
+        <ul className='ms-0 flex flex-col ps-0 before:hidden'>
+          {sortByOptions.map((option) => (
+            <MenuItem
+              key={option.value}
+              label={option.label}
+              buttonClass='h-8'
+              toggled={sortBy === option.value}
+              onClick={() => handleSetSortBy(option.value as LibrarySortByType)}
+              transient
+            />
+          ))}
+          <hr aria-hidden='true' className='border-base-200 my-1' />
+          {sortingOptions.map((option) => (
+            <MenuItem
+              key={option.value.toString()}
+              label={option.label}
+              buttonClass='h-8'
+              toggled={isAscending === option.value}
+              onClick={() => handleSetSortAscending(option.value)}
+              transient
+            />
+          ))}
+        </ul>
+      </MenuItem>
     </Menu>
   );
 };
